[Code]: 데이터 파일 생성
