+++ W1Team5/W1Team5/HamsikInfo.swift
@@ -1,0 +1,21 @@
+//
+//  HamsikInfo.swift
+//  W1Team5
+//
+//  Created by 황석현 on 10/24/24.
+//
+
+import Foundation
+
+// MARK: 회의에서 정한 데이터 구조
+struct HamsikInfo {
+    let name: String
+    let description: String
+    let mbti: String
+    let strengths: [String]
+    let style: [String]
+    let cooperation: [String]
+    let blogURL: String
+    let profileImage: String
+    let zepImage: String
+}
